Updated get_move() method
Instead of returning a Move item, it will now return a pair of optional Move and optional InvalidInput objects.

This will be good as now all the Player objects will no longer be handling loops for getting a valid input and also printing stuff to the console.

This commit also comes with a bunch of documentation for some of the Player subclasses.