Fixed bug involving InvalidInput return type
In my implementation I thought that this would work with subclasses of InvalidInput to specify more precise invalid types of inputs using run-time polymorphism.

However, that is not the case as it needs to be a pointer or it will always run the message() method of the InvalidInput scoped as the InvalidInput's method and not the subclasses' overridden method. So, I made InvalidInput into a pointer, and since it is a pointer, I don't need it to be optional as I can just check for 'nullptr' instead of 'nullopt' for optional.